* Add an option to `bitten-slave` that tells it to keep any files in the working directory.  * Some safety measures for the unpacking of `tar` archives: no `chown`-ing, no absolute paths or paths containing up-references (`..`).
git-svn-id: http://svn.edgewall.org/repos/bitten/trunk@244 2d6df8db-84f7-0310-ba0f-a4c632ebc3b5

+++ bitten/slave.py
@@ -31,7 +31,7 @@ class Slave(beep.Initiator):
     """Build slave."""
 
     def __init__(self, ip, port, name=None, config=None, dry_run=False,
-                 work_dir=None):
+                 work_dir=None, keep_files=False):
         beep.Initiator.__init__(self, ip, port)
         self.name = name
         self.config = config
@@ -41,6 +41,7 @@ class Slave(beep.Initiator):
         elif not os.path.exists(work_dir):
             os.makedirs(work_dir)
         self.work_dir = work_dir
+        self.keep_files = keep_files
 
     def greeting_received(self, profiles):
         if OrchestrationProfileHandler.URI not in profiles:
@@ -150,8 +151,9 @@ class OrchestrationProfileHandler(beep.ProfileHandler):
                 recipe = Recipe(self.recipe_xml, path, self.config)
                 self.execute_build(msgno, recipe)
             finally:
-                shutil.rmtree(path)
-                os.unlink(archive_path)
+                if not self.session.keep_files:
+                    shutil.rmtree(path)
+                    os.remove(archive_path)
 
     def execute_build(self, msgno, recipe):
         log.info('Building in directory %s', recipe.ctxt.basedir)
@@ -244,6 +246,9 @@ def main():
                       metavar='FILE', help='path to configuration file')
     parser.add_option('-d', '--work-dir', action='store', dest='work_dir',
                       metavar='DIR', help='working directory for builds')
+    parser.add_option('-k', '--keep-files', action='store_const',
+                      dest='keep_files', const=True,
+                      help='don\'t delete files after builds')
     parser.add_option('-n', '--dry-run', action='store_const', dest='dry_run',
                       const=True, help='don\'t report results back to master')
     parser.add_option('--debug', action='store_const', dest='loglevel',
@@ -276,13 +281,14 @@ def main():
     logger.addHandler(handler)
 
     slave = Slave(host, port, name=options.name, config=options.config,
-                  dry_run=options.dry_run, work_dir=options.work_dir)
+                  dry_run=options.dry_run, work_dir=options.work_dir,
+                  keep_files=options.keep_files)
     try:
         slave.run()
     except KeyboardInterrupt:
         slave.quit()
 
-    if os.path.isdir(slave.work_dir):
+    if not options.keep_files and os.path.isdir(slave.work_dir):
         shutil.rmtree(slave.work_dir)
 
 if __name__ == '__main__':
